// src/components/Battle.module.css.d.ts
declare const styles: {
  arena: string;
  arenaFloor: string;
  character: string;
  characterLeft: string;
  characterRight: string;
  blocking: string;
  stunned: string;
  characterModel: string;
  playerModel: string;
  opponentModel: string;
  arenaLighting: string;
};

export default styles; 